# add postgresql index and use IN instead of many OR