--- a/quickwit/quickwit-metastore/migrations/postgresql/16_create-index-split-uid.down.sql
+++ b/quickwit/quickwit-metastore/migrations/postgresql/16_create-index-split-uid.down.sql
@@ -1,0 +1,1 @@
+DROP INDEX IF EXISTS index_splits_index_uid;