--- a/quickwit/quickwit-metastore/migrations/postgresql/16_create-index-split-uid.up.sql
+++ b/quickwit/quickwit-metastore/migrations/postgresql/16_create-index-split-uid.up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS index_splits_index_uid ON splits USING HASH(index_uid);