### Description

fix part of #4658 (but we'll keep that open to explore other indexes we may want to add)
add an index on `splits(index_uid)`

fix #4666
use `index_uid IN (<list of index>)` instead of `index_uid = "index1" OR index_uid = "index2" OR ...` when querying multiple index.

### How was this PR tested?

tested by measuring time and checking query plan on an index with a few hundred thousand splits.
`SELECT * FROM "splits" WHERE "index_uid" = $1` is made ~1900x faster on that particular dataset using the index (from 5.8s to 3.1ms)
selecting with a list of 1k indexes, covering almost every split, is made ~3.5x faster (from 3s to 850ms) by using `IN` over many `OR`. Execution time and query plan are not affected by `IN` over a single `=` when searching with a single index